Allow attenuation for `sensor.adc` to be null
Currently it isn't possible to define an ADC sensor in a package that is intended to be used on both ESP8266 and ESP32 because there is no null value for the `attenuation` key and no way to remove the key from the configuration.

This problem isn't unique to `sensor.adc` and would benefit from a generalized solution. The best option I could think of was to filter out null values in `esphome.config.validate_config`, but that would prevent components from accepting a `null` value altogether, which may not be desirable.